약관 등록 메서드 어드민 컨트롤러 생성후 이동
